Allow for navigation when deciding on words to save, improve word decision experience
Moving back and forth is possible with left and right arrows. Navigation is unified between proficiency selection and browsing of words.
The choice made on a word is briefly shown before moving to the next word.

On the first showing of the word prompt, no changes to the word index are made. This is why the word prompt subroutine now gets a parameter 'advance' to say whether to move at all. The word index is now incremented at the beginning of the word prompt subroutine, because we want to decide how much to move to get to the next word, not the word after it. This is ultimately because we now do not know which word that will be, as the user has more choice.